+++ Applications/Window/Plot.cpp
@@ -545,14 +545,14 @@ void As::Plot::updateGraphOnPlot(const QPair<QVector<int>, QVector<int> > ranges
 */
 void As::Plot::updateAllOnPlot(const As::Scan *scan)
 {
-    ADEBUG << scan;
+    //ADEBUG << scan;
     // Get data to plot
     As::RealVector x  = (*scan)["angles"][scan->scanAngle()]["data"];
     As::RealVector y  = (*scan)["intensities"]["DetectorNorm"]["data"];
     As::RealVector sy = (*scan)["intensities"]["sDetectorNorm"]["data"];
-    ADEBUG << x;
-    ADEBUG << y;
-    ADEBUG << sy;
+    //ADEBUG << x;
+    //ADEBUG << y;
+    //ADEBUG << sy;
     // Update axes ranges
     updateAxesRanges(x, y, sy); // Auto by QCustomPlot: rescaleAxes();
     // Remove all existing graphs